New function Test-ScriptVersion implemented
This function checks if we're connected to the internet (in this case we query api.github.com) and after that we determine if the version we're running is the latest version. If no internet connection is possible the script continues but we suggest to check manually for the latest version.