Get-ReplacingServerUpdates improvements
To address issue #250 . We've included -UseBasicParsing which is default with PowerShell 6. This change requires to update the parsing logic to make sure to get the right Microsoft Update Catalog guid.
Changed also Update Catalog url from http to https and added Stop-Job before Remove-Job if we reach the specified Timeout value.